make RecordScreen.kt
